requirements.txt
Pillow==8.4.0
tkinter